+++ src/main/java/com/logpie/shopping/tool/controller/ClientController.java
@@ -9,13 +9,15 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.logpie.framework.db.basic.Page;
+import com.logpie.dba.api.basic.Page;
 import com.logpie.framework.log.util.LogpieLogger;
 import com.logpie.framework.log.util.LogpieLoggerFactory;
 import com.logpie.shopping.tool.model.Client;
 import com.logpie.shopping.tool.model.Shop;
 import com.logpie.shopping.tool.service.ClientService;
 import com.logpie.shopping.tool.service.ShopService;
+
+import java.math.BigInteger;
 
 @Controller
 @RequestMapping("/{shopPath}/clients")
@@ -48,7 +50,7 @@ public class ClientController {
 	}
 
 	@RequestMapping(path = "/id/{id}", method = RequestMethod.GET)
-	public @ResponseBody Client getByAJAX(@PathVariable final Long id) {
+	public @ResponseBody Client getByAJAX(@PathVariable final BigInteger id) {
 		return service.getClientById(id);
 	}
