+++ src/main/java/com/logpie/shopping/tool/model/Express.java
@@ -1,21 +1,23 @@
 package com.logpie.shopping.tool.model;
 
-import com.logpie.framework.db.annotation.AutoGenerate;
-import com.logpie.framework.db.annotation.AutoGenerate.AutoGenerateType;
-import com.logpie.framework.db.annotation.Column;
-import com.logpie.framework.db.annotation.Column.DataType;
-import com.logpie.framework.db.annotation.ForeignKeyColumn;
-import com.logpie.framework.db.annotation.ID;
-import com.logpie.framework.db.annotation.Entity;
-import com.logpie.framework.db.basic.Model;
+import com.logpie.dba.api.annotation.Entity;
+import com.logpie.dba.api.annotation.ID;
+import com.logpie.dba.api.annotation.ForeignKeyColumn;
+import com.logpie.dba.api.annotation.Column;
+import com.logpie.dba.api.annotation.Column.DataType;
+import com.logpie.dba.api.annotation.AutoGenerate;
+import com.logpie.dba.api.annotation.AutoGenerate.AutoGenerateType;
+import com.logpie.dba.api.basic.Model;
 import com.logpie.shopping.tool.repository.ExpressRepository;
+
+import java.math.BigInteger;
 
 @Entity(name = ExpressRepository.DB_TABLE_EXPRESS)
 public class Express extends Model {
 	@ID
 	@AutoGenerate(strategy = AutoGenerateType.NumberAutoIncrement)
-	@Column(name = ExpressRepository.DB_KEY_EXPRESS_ID, type = DataType.LONG)
-	private Long id;
+	@Column(name = ExpressRepository.DB_KEY_EXPRESS_ID, type = DataType.BIGINT)
+	private BigInteger id;
 
 	@Column(name = ExpressRepository.DB_KEY_EXPRESS_NAME, type = DataType.STRING)
 	private String name;
@@ -37,15 +39,15 @@ public class Express extends Model {
 	 * @param isInternational
 	 * @param shop
 	 */
-	public Express(final Long id, final String name,
-			final Boolean isInternational, final Shop shop) {
+	public Express(final BigInteger id, final String name,
+				   final Boolean isInternational, final Shop shop) {
 		this.id = id;
 		this.name = name;
 		this.isInternational = isInternational;
 		this.shop = shop;
 	}
 
-	public Long getId() {
+	public BigInteger getId() {
 		return id;
 	}
 
@@ -61,7 +63,7 @@ public class Express extends Model {
 		return shop;
 	}
 
-	public void setId(Long id) {
+	public void setId(BigInteger id) {
 		this.id = id;
 	}
