+++ src/main/java/com/logpie/shopping/tool/model/ShippingRecord.java
@@ -1,21 +1,23 @@
 package com.logpie.shopping.tool.model;
 
-import com.logpie.framework.db.annotation.AutoGenerate;
-import com.logpie.framework.db.annotation.AutoGenerate.AutoGenerateType;
-import com.logpie.framework.db.annotation.Column;
-import com.logpie.framework.db.annotation.Column.DataType;
-import com.logpie.framework.db.annotation.Entity;
-import com.logpie.framework.db.annotation.ForeignKeyColumn;
-import com.logpie.framework.db.annotation.ID;
-import com.logpie.framework.db.basic.Model;
+import com.logpie.dba.api.annotation.Entity;
+import com.logpie.dba.api.annotation.ID;
+import com.logpie.dba.api.annotation.ForeignKeyColumn;
+import com.logpie.dba.api.annotation.Column;
+import com.logpie.dba.api.annotation.Column.DataType;
+import com.logpie.dba.api.annotation.AutoGenerate;
+import com.logpie.dba.api.annotation.AutoGenerate.AutoGenerateType;
+import com.logpie.dba.api.basic.Model;
 import com.logpie.shopping.tool.repository.ShippingRecordRepository;
+
+import java.math.BigInteger;
 
 @Entity(name = ShippingRecordRepository.DB_TABLE_SHIPPING_RECORD)
 public class ShippingRecord extends Model {
 	@ID
 	@AutoGenerate(strategy = AutoGenerateType.NumberAutoIncrement)
-	@Column(name = ShippingRecordRepository.DB_KEY_SHIPPING_RECORD_ID, type = DataType.LONG)
-	private Long id;
+	@Column(name = ShippingRecordRepository.DB_KEY_SHIPPING_RECORD_ID, type = DataType.BIGINT)
+	private BigInteger id;
 
 	@ForeignKeyColumn(name = ShippingRecordRepository.DB_KEY_SHIPPING_RECORD_PACKAGE_ID, referencedEntityClass = Package.class)
 	private Package pack;
@@ -24,17 +26,17 @@ public class ShippingRecord extends Model {
 	private Transaction transaction;
 
 	public ShippingRecord() {
-
+		
 	}
 
-	public ShippingRecord(final Long id, final Package pack,
+	public ShippingRecord(final BigInteger id, final Package pack,
 			final Transaction transaction) {
 		this.id = id;
 		this.pack = pack;
 		this.transaction = transaction;
 	}
 
-	public Long getId() {
+	public BigInteger getId() {
 		return id;
 	}
 
@@ -46,7 +48,7 @@ public class ShippingRecord extends Model {
 		return transaction;
 	}
 
-	public void setId(Long id) {
+	public void setId(BigInteger id) {
 		this.id = id;
 	}
