+++ src/main/java/com/logpie/shopping/tool/repository/ProductConfigRepository.java
@@ -1,5 +1,6 @@
 package com.logpie.shopping.tool.repository;
 
+import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
@@ -8,9 +9,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Repository;
 
-import com.logpie.framework.db.basic.Parameter;
-import com.logpie.framework.db.basic.WhereParam;
-import com.logpie.framework.db.repository.JDBCTemplateRepository;
+import com.logpie.dba.api.basic.Parameter;
+import com.logpie.dba.api.basic.WhereParam;
+import com.logpie.dba.api.repository.JDBCTemplateRepository;
 import com.logpie.shopping.tool.model.ProductConfig;
 
 @Repository
@@ -33,7 +34,7 @@ public class ProductConfigRepository extends
 		super(ProductConfig.class);
 	}
 
-	public List<ProductConfig> queryByProductId(final Long productId)
+	public List<ProductConfig> queryByProductId(final BigInteger productId)
 			throws DataAccessException {
 		Parameter param = new WhereParam(ProductConfig.class,
 				DB_KEY_CONFIG_PRODUCT_ID, productId);
